backend/collegeBE/requirements.txt
amqp==2.5.2
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
billiard==3.6.2.0
celery==4.4.0
Django==1.11
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
importlib-metadata==1.5.0
isort==4.3.21
joblib==0.14.1
jsonschema==3.2.0
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyrsistent==0.15.7
pytz==2019.3
redis==3.4.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
typed-ast==1.4.1
vine==1.3.0
wrapt==1.11.2
zipp==2.2.0
backcall==0.1.0
decorator==4.4.2
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
numpy==1.18.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickle5==0.0.9
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
